Moved computation store from api into store folder
