+++ src/components/interpreter.tsx
@@ -6,16 +6,12 @@ import { javascript } from "@codemirror/lang-javascript";
 import CodeMirror, { ReactCodeMirrorRef } from "@uiw/react-codemirror";
 import beautify from "js-beautify";
 
-import { computationStore } from "../api/computation";
-import {
-  refresh,
-  stepForward,
-  stepToIndex,
-} from "../engine/manual/execute";
+import { executionStore as ctx } from "../api/execution";
+import { refresh, stepForward, stepToIndex } from "../engine/manual/execute";
 import { extractManual } from "../engine/manual/extract";
 import { isAtBlock } from "../engine/manual/interpreter";
 import { Step } from "../engine/manual/step";
-import { executionStore as ctx } from "../api/execution";
+import { computationStore } from "../store/computation";
 import { IEnvironment } from "../types/environment";
 import { extractViews } from "../util/acorn";
 import {
