+++ src/visualizations/plot3d/Plot3D.tsx
@@ -14,11 +14,11 @@ import * as Plotly from "plotly.js-dist";
 
 import { IPlot3D } from "../../api";
 import { FormulizeConfig } from "../../api";
-import { computationStore } from "../../api/computation";
 import {
   computeSurfaceIntersection,
   solveSingularFormula,
 } from "../../engine/singular-formula-solver";
+import { computationStore } from "../../store/computation";
 import { IPoint3D, ISurface } from "../../types/plot3d";
 import { getVariable, getVariableValue } from "../../util/computation-helpers";
 import { getFormulaByName } from "../../util/formula-by-name";
